# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test Application",
        "version": "0.1"
    },
    "paths": {
...
        "/api/v1/ComputerForProgram": {
            "get": {
                "tags": [
                    "ComputerForProgram"
                ],
                "summary": "Get a list of Computer For Program",
                "description": "",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "ComputerForProgram"
                ],
                "summary": "Create new prgram for computer",
                "description": "",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "computer_id": {
                                        "description": "computer_id",
                                        "type": "number"
                                    },
                                    "program_id": {
                                        "description": "program_id",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Validate error"
                    },
                    "200": {
                        "description": "Success create element"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/ComputerForProgram/{id}": {
            "get": {
                "tags": [
                    "ComputerForProgram"
                ],
                "summary": "Get one program by computer id",
                "description": "",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Computer id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "ComputerForProgram"
                ],
                "summary": "Updates a ComputerForProgram in the store with form data",
                "description": "",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of ComputerForProgram that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "computer_id": {
                                        "description": "computer_id",
                                        "type": "number"
                                    },
                                    "program_id": {
                                        "description": "program_id",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "500": {
                        "description": "Error in server"
                    },
                    "400": {
                        "description": "Validate error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/computerManufactory": {
            "get": {
                "tags": [
                    "Computer Monofaktura"
                ],
                "summary": "Get a list of computer Manufactory",
                "description": "",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Computer Monofaktura"
                ],
                "summary": "Create new Monofaktura for computer",
                "description": "",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "name",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Validate error"
                    },
                    "200": {
                        "description": "Success create element"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/computerManufactory/{id}": {
            "get": {
                "tags": [
                    "Computer Monofaktura"
                ],
                "summary": "Get  Computer Monofaktura by id",
                "description": "",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Computer Monofaktura"
                ],
                "summary": "Updates a Computer Monofaktura in the store with form data",
                "description": "",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Computer Monofaktura that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "name",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "500": {
                        "description": "Error in server"
                    },
                    "400": {
                        "description": "Validate error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}